Ratios between pulse frequencies and prime numbers.
Pulse frequencies 78250 Hz and 31300 Hz. Memo: 78250/31300 = 5/2.

78250 / (5 * 3 * 11 * 13 * 19) = 78250/40755 = 1.92
31300 / (2 * 3 * 11 * 13 * 19) = 31300/16302 = 1.92

where 1.92 are 96 cycles at 50 Hz, the repetition period of the signal
at 422.733 MHz.

Maybe the infrasound is generated by the pulses (with offset from the
received radiation) induced into the radio module of the smart meter.
In this case, it is also possible to exploit these pulses to control
the radio module, because the received radiation is the offset of the
pulse train, therefore the result is an AM demodulated train of impulse
responses (from BP filter with fcut=169MHz), where the value of a IR
peak is proportional to the modulation of the received radiation
(it is a scaled sample value of that modulation).