Cuba: pulses from electrical grid with frequency 77940 or 56340 Hz ?
If the system is the same:

  # Italy, nominal frequency 50 Hz and magic ratio 5/2
  78250 = 313 * 250
  31300 = 313 * 100
  # Cuba, nominal frequency 60 Hz and magic ratio 3/2 (or 3)
  56340 = 313 * 180
  37560 = 313 * 120
  18780 = 313 * 60

313 is a prime number.

The alternative with the prime number 433 and magic ratio 3/2 or 3:

  77940 = 433 * 60 * 3 = 433 * 120 * 3/2    [near 78250 Hz]
  51960 = 433 * 120
  25980 = 433 * 60

Example: multiple of 180 Hz from aliasing of pulses with pulse
frequency 1800=5400/3 Hz (instead of 1733=5200/3 Hz here) starting
from 943.695e6 MHz. Infrasounds generated from

  253.5 MHz = 169 MHz * 3/2, if the magic ratio is 3/2
  or 507 MHz = 169 MHz * 3, if the magic ratio is 3

All the multiple of 180 Hz from 180 to 28080 by aliasing of
157 sinusoids separated by 180 * N, with N positive integer
and sampling rate 56340 Hz. [157 = (56340/180 + 1)/2]