Audio recordings: 100 Hz, again.
Normalized and LP filtered.

Meanwhile I'm investigating hypnagogic and hypnopompic hallucinations
(there is not a night without these fake dreams) from machine learning
with a very bad virtual reality, often a few dark frames for second
(also isolated images). I recognize the style, the culture and the
intent of the attacker. 98% of the images (75% of the faces) are new
for me (however some themes are repeated with little variations).
The short "films" always terminate with a highly emotive scene to
force an awakening (sometimes it is the intent but I don't feel
emotions; in this case the scene terminates after a few seconds of
boring frames). Sometimes there is a loop during the hypnopompic state
(i.e. chewing food for about one or two seconds for cycle) and the
awakening is delayed. It is necessary to do other experiments before
to publish some tricks. There are at least two possibilities: circular
(or spiral) rastering with laser on retina (I'm not expert at moment)
and some type of brain clock/interrupt, or the interesting technique
deduced from the following article:

  https://gizmodo.com/scientists-reconstruct-brains-visions-into-digital-vide-5843117

The other techniques to publish regard voices/sounds from laser,
arbitrary-source/voice-formants decomposition and spectral effects
with directional cues. The frightened idiots stopped to use these
tecniques the last year, when I started to write this document but I
remember every steps with details.