Cancel the information about the perforated aluminum foils.
This clue is weak. I can see very little holes with a lamp behind the
other foils. Perhaps I could consider the density of the holes in
different parts of the window.

I don't know if this experiment is stupid, however it will remain in
the history of this document.

The two little holes in the aluminium container is a particular clue.