Strange circular patterns on a CD inserted in the right pocket.
I think the circular modulation is the key of that weapon.

I think it is also important the radius modulation and the amplitude
modulation of these geometries. Probably the combination of these
modulations determine a continuation from pressure wave (photoacoustic
spectroscopy) to air movement (dynamical spatial distribution of the
vibrations, for example the tangible sensation of an invisible water
cannon -- the last aborted with some CDs on July 24 2019 night,
between 2:00 and 2:30 [*]).

[*] Is it similar to the weapon named "anvil" announced by a English
    general (MIR if my memory is correct)?
    A coincidence about the anvil: during the first days of the attack
    (2017), the worm-team pointed the pulsed radiation into my right
    ear, and after the usual chorus (the vocal "A" from multiple
    artificial voices, with repeated glissando up and down, a kind of
    aliasing around 4 or 5 kHz), an aggressive (for psychological impact)
    sampled Italian voice blethered: "La senti l'incudine? La senti
    l'incudine?" ("Do you hear the anvil? Do you hear the anvil?").
    I don't know if there is also an allusion to my software "incudine".