Mobile communications (any generation).
An emollient ointment based on glycerine (about 4mm thick) on the head
is good to attenuate the tinnitus. My sleep deprivation is reduced.
The fake dreams (external images on retina from a kind of GPT) remain
but I know how to control the psychological impact. Note: that film
doesn't work without a storyteller (hypnosis from voice) and my
anomalous breath modulation is intense when the fake dream includes
music and sound effects (I think that modulation is related to teeth
and jaw vibration).

There are moles on the head where the tinnitus is stimulated (variable
location, 99% left side). I don't know if it is a coincidence.